+++ packages/nuxt/build.config.ts
@@ -7,22 +7,35 @@ const external = [
   // @ts-expect-error
   ...Object.keys(pkg.peerDependencies || {}),
 ]
-export default defineBuildConfig({
-  entries: [
-    { input: 'src/core/', outDir: 'dist/core' },
-    { input: 'src/modules/', outDir: 'dist/modules' },
-  ],
-  // explicitly externalize consola since Nuxt has it
-  externals: [
-    'consola',
-    '@pergel/graphql',
-    '@apollo/sandbox',
-    '#pergel',
-    'h3',
-    'pathe',
-    'scule',
-    'unimport',
-    'defu',
-    ...external,
-  ],
-})
+export default defineBuildConfig([
+  {
+    entries: [
+      {
+        input: 'src/core/',
+        outDir: 'dist/core',
+        ext: 'mjs',
+      },
+    ],
+    rollup: {
+      esbuild: {
+        target: 'esnext',
+      },
+      emitCJS: false,
+      cjsBridge: true,
+    },
+
+    // explicitly externalize consola since Nuxt has it
+    externals: [
+      'consola',
+      '@pergel/graphql',
+      '@apollo/sandbox',
+      '#pergel',
+      'h3',
+      'pathe',
+      'scule',
+      'unimport',
+      'defu',
+      ...external,
+    ],
+  },
+])
