+++ packages/nuxt/client/composables/state.ts
@@ -1,4 +1,4 @@
 import { ref } from 'vue'
-import type { PergelOptions } from '../../src/core/types'
+import type { PergelOptions } from '../../src/runtime/core/types'
 
 export const options = ref<PergelOptions>()
