+++ packages/nuxt/package.json
@@ -51,7 +51,9 @@
     "@nuxt/module-builder": "^0.5.4",
     "@nuxt/schema": "^3.8.2",
     "@nuxt/test-utils": "^3.8.1",
+    "@rollup/plugin-node-resolve": "^15.2.3",
     "@types/node": "^20.9.3",
+    "esbuild-plugin-file-path-extensions": "^2.0.0",
     "nuxt": "^3.8.2",
     "vitest": "^0.33.0"
   },
