+++ packages/nuxt/playground/pergel/README.yaml
@@ -1,15 +1,15 @@
 pergel:
-# This file is generated by pergel. Do not edit it manually.
-# Version: 0.0.1-beta.4
+  # This file is generated by pergel. Do not edit it manually.
+  # Version: 0.0.1-beta.5
 test:
   S3:
     env:
       NUXT_TEST_S3_REGION: auto
-      NUXT_TEST_S3_ENDPOINT:
-      NUXT_TEST_S3_ACCESS_KEY_ID:
-      NUXT_TEST_S3_SECRET_ACCESS_KEY:
-      NUXT_TEST_S3_BUCKET:
+      NUXT_TEST_S3_ENDPOINT: 
+      NUXT_TEST_S3_ACCESS_KEY_ID: 
+      NUXT_TEST_S3_SECRET_ACCESS_KEY: 
+      NUXT_TEST_S3_BUCKET: 
     # If pergel cli is installed, you can run `pergel install` automatically to install
     packageJson:
-      dependencies: '@pergel/module-s3@0.0.0'
-      devDependencies: ''
+      dependencies: "@pergel/module-s3@0.0.0"
+      devDependencies: ""
