+++ packages/nuxt/src/runtime/core/constants/index.ts
