+++ packages/nuxt/src/runtime/core/types/devtools.ts
