+++ packages/nuxt/src/runtime/core/types/module.ts
