+++ packages/nuxt/src/runtime/core/utils/checkOptions.ts
