+++ packages/nuxt/src/runtime/core/utils/generateYaml.ts
