+++ packages/nuxt/src/runtime/core/utils/pergelFunctionTemplate.ts
