+++ packages/nuxt/src/runtime/core/utils/rootFolderSync.ts
